+++ frontend/src/components/common/CopyTextButton/styles.ts
