+++ frontend/src/components/common/index.tsx
@@ -14,3 +14,5 @@ export { default as EmptyContent } from './EmptyContent';
 export { default as OptionSwitch } from './OptionSwitch';
 export { default as ReviewEmptySection } from './ReviewEmptySection';
 export * from './modals';
+export { default as CopyTextButton } from './CopyTextButton';
+export { default as ReviewCard } from './ReviewCard';
