+++ frontend/src/components/index.tsx
@@ -3,4 +3,5 @@ export * from './common';
 export * from './error';
 export * from './highlight/components';
 export * from './login';
+export * from './reviewURL';
 export * from './skeleton';
