+++ frontend/src/components/reviewURL/URLGeneratorForm/hooks/usePostDataForReviewRequestCode/test.tsx
@@ -1,6 +1,6 @@
 import { renderHook, act, waitFor } from '@testing-library/react';
 
-import { CREATED_REVIEW_REQUEST_CODE } from '@/mocks/mockData/group';
+import { VALID_REVIEW_GROUP_REVIEW_REQUEST_CODE } from '@/mocks/mockData/group';
 import QueryClientWrapper from '@/queryTestSetup/QueryClientWrapper';
 
 import usePostDataForReviewRequestCode from '.';
@@ -33,6 +33,6 @@ describe('usePostDataForReviewRequestCode', () => {
     await waitFor(() => expect(result.current.isSuccess).toBe(true));
 
     // then
-    expect(result.current.data).toEqual(CREATED_REVIEW_REQUEST_CODE);
+    expect(result.current.data.reviewRequestCode).toEqual(VALID_REVIEW_GROUP_REVIEW_REQUEST_CODE);
   });
 });
