+++ frontend/src/components/reviewURL/index.ts
@@ -1,0 +1,2 @@
+export { default as URLGeneratorForm } from './URLGeneratorForm';
+export { default as ReviewZoneURLModal } from './ReviewZoneURLModal';
