--- frontend/src/pages/HomePage/components/FormBody/styles.ts
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-export const FormBody = styled.div<{ direction: React.CSSProperties['flexDirection'] }>`
-  display: flex;
-  flex-direction: ${({ direction }) => direction};
-  gap: 1.8rem;
-`;
