--- frontend/src/pages/HomePage/components/FormLayout/styles.ts
@@ -1,13 +1,0 @@
-import styled from '@emotion/styled';
-
-export const FormLayout = styled.form`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-`;
-
-export const Title = styled.h2`
-  margin-bottom: 6.5rem;
-  font-size: 2.5rem;
-  white-space: nowrap;
-`;
