--- frontend/src/pages/HomePage/components/index.ts
@@ -1,7 +1,0 @@
-export { default as FormBody } from './FormBody';
-export { default as FormLayout } from './FormLayout';
-export { default as ReviewZoneURLModal } from './ReviewZoneURLModal';
-export { default as URLGeneratorForm } from './URLGeneratorForm';
-export { default as CopyTextButton } from './CopyTextButton';
-export { default as ReviewMeOverview } from './ReviewMeOverview';
-export { default as InfinityCarousel } from './InfinityCarousel';
