+++ frontend/src/pages/HomePage/components/index.tsx
@@ -1,0 +1,2 @@
+export { default as ReviewMeOverview } from './ReviewMeOverview';
+export { default as InfinityCarousel } from './InfinityCarousel';
