+++ frontend/src/utils/index.ts
@@ -12,3 +12,4 @@ export * from './media';
 export * from './highlight/index';
 export * from './testUtils';
 export * from './analytics';
+export * from './validateInput';
